fsds_cleaning_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fsds-cleaning-env"
version = "0.1.0"
description = "HF-ready OpenEnv environment for the data-cleaning stage of Full-Stack Data Science"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.1",
  "fastapi>=0.115.0",
  "fastmcp>=2.0.0",
  "pydantic>=2.0.0",
  "uvicorn>=0.24.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
]

[project.scripts]
server = "fsds_cleaning_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["fsds_cleaning_env", "fsds_cleaning_env.server", "fsds_cleaning_env.training"]
package-dir = { "fsds_cleaning_env" = ".", "fsds_cleaning_env.server" = "server", "fsds_cleaning_env.training" = "training" }
